Adds description and resources related to the service distributions
- Added few minor changes to the repository structure, mainly related to
the `releases` directory.
  It will contain information about our internal applications for the
reconciliation. Currently there will be 2 files, one for the
`Reconciliator` and one for the `Index Manager`. The main idea is to
sync these files with our internal release documentation so that we keep
the users up-to-date with the changes in different versions.

- Added official documentation and resources for Docker distribution.

- Added official documentation and resources for Helm (k8s)
distribution.

Each of the distributions contains examples how to setup the environment
and what configurations are supported (also required) for the different
components/services.

- Removed the ToC from the `documentation/README.md` file. GitHub
generates it in a side bar, which is a lot more convenient.