Adds updates to the distributions and their documentation
- Added `.env` file for the Docker deployment. The file contains some
common arguments, which should help while updating to newer versions or
manager the resources for the services.

- Added several changes to the Helm deployment. The aim is to make the
deployment of the services more fluent and at the same time to ensure
that all of the resources and files are provided.

- There were few QoL updates to the Index Manager and Reconciliator
applications in order to make them more compatible for deployment in
Kubernetes cluster.

- Added few examples of a simple setup for Helm deployment.

- Added few changes in the deployments documentation.

- Added few updates to the `.gitignore` and `.helmignore` files.